// src/components/button/ButtonConstants.d.ts
import { ButtonSize } from './types';
export declare const PADDINGS: {
    XSMALL: number;
    SMALL: number;
    MEDIUM: number;
    LARGE: number;
};
export declare const HORIZONTAL_PADDINGS: {
    XSMALL: number;
    SMALL: number;
    MEDIUM: number;
    LARGE: number;
};
export declare const MIN_WIDTH: {
    XSMALL: number;
    SMALL: number;
    MEDIUM: number;
    LARGE: number;
};
export declare const SIZE_TO_VERTICAL_HITSLOP: {
    readonly xSmall: 30;
    readonly small: 25;
    readonly medium: 20;
    readonly large: 15;
};
export declare const DEFAULT_SIZE = ButtonSize.large;